pyproject.toml
[project]
name = "quantagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "anthropic>=0.88.0",
    "flask>=3.1.3",
    "ipython>=9.10.1",
    "langchain>=1.2.14",
    "langchain-anthropic>=1.4.0",
    "langchain-core>=1.2.24",
    "langchain-openai>=1.1.12",
    "langchain-qwq>=0.3.4",
    "langgraph>=1.1.4",
    "matplotlib>=3.10.8",
    "mcp[cli]>=1.27.0",
    "mplfinance>=0.12.10b0",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scipy>=1.17.1",
    "ta-lib>=0.6.8",
    "yfinance>=1.2.0",
]
